Update protonvpn-cli to 3.11.0
Summary:
Depends on D12679, D12680.

CHANGELOG available [here](https://github.com/ProtonVPN/linux-cli/releases/tag/3.11.0).

This is also a refactoring of the package (particularly separating `proton-python-client` and `protonvpn-nm-lib`) so that it will be easier to manage the dependencies since the inclusion of `protonvpn-gui` (T9924).

Test Plan:
Connected to various locations in the world, checked IP addresses and ran speed tests.

Signed-off-by: Gavin Zhao <git@gzgz.dev>

Reviewers: #triage_team, Girtablulu

Reviewed By: #triage_team, Girtablulu

Subscribers: Girtablulu

Differential Revision: https://dev.getsol.us/D12681